# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Music Room API",
    "version": "1.0.0"
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "paths": {
...
    "/api/v1/playlists/{id}/reorder": {
      "patch": {
        "responses": {
          "200": {
            "description": "{version}"
          },
          "409": {
            "description": "stale version + current order"
          }
        },
        "summary": "Versioned reorder"
      }
    },
...
    "/api/v1/suggestions/{id}/vote": {
      "post": {
        "responses": {
          "200": {
            "description": "voted"
          },
          "403": {
            "description": "license/geofence denied"
          },
          "409": {
            "description": "already voted"
          }
        },
        "summary": "Atomic vote"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}